cleanup

+++ index.js
@@ -1,18 +1,21 @@
 var q = require('q');
+var fs = require('fs');
 var spawn = require('child_process').spawn;
 var whichSync = require('which').sync;
 var Xvfb = require('xvfb');
 var chromeLocation = require('chrome-location');
 
-var HeadlessChromeBrowser = function(baseBrowserDecorator, args) {
+var HeadlessChromeBrowser = function (baseBrowserDecorator, config, emitter, args) {
 
   var self = this;
-  baseBrowserDecorator(this);
 
   this.name = 'HeadlessChrome';
 
+  baseBrowserDecorator(self);
+
   var xvfb = new Xvfb({
     silent: true,
+    reuse: true,
     xvfb_args: [
       '-screen',
       '0',
@@ -21,21 +24,26 @@ var HeadlessChromeBrowser = function(baseBrowserDecorator, args) {
     ]
   });
 
+  fs.openSync(xvfb._lockFile(), 'a');
+  var display = xvfb.display();
+
   var chromeProcess;
 
   this.start = function (url) {
 
     if (!chromeLocation) {
       console.error('Cannot find Chrome executable.');
-      self.kill();
+      self.forceKill();
     }
 
     try {
       whichSync('Xvfb');
     } catch (e) {
       console.error('Cannot find Xvfb executable.');
-      self.kill();
+      self.forceKill();
     }
+
+    var display = xvfb.display();
 
     var chromeFlags = [
       '--user-data-dir=' + this._tempDir,
@@ -44,13 +52,14 @@ var HeadlessChromeBrowser = function(baseBrowserDecorator, args) {
       '--disable-default-apps',
       '--disable-popup-blocking',
       '--disable-translate',
-      '--disable-background-timer-throttling'
+      '--disable-background-timer-throttling',
+      '--display=' + display
     ].concat(args.chromeFlags || [], [url]);
 
-    xvfb.start(function(err) {
+    xvfb.start(function (err) {
       if (err) {
         console.error('Failed to start Xvfb: ' + err);
-        self.kill();
+        self.forceKill();
         return;
       }
       chromeProcess = spawn(chromeLocation, chromeFlags);
@@ -58,25 +67,24 @@ var HeadlessChromeBrowser = function(baseBrowserDecorator, args) {
 
   };
 
-  this.kill = function (done) {
+  this.on('kill', function (done) {
     function allDone() {
       self._done();
       if (done) {
         done();
       }
     }
+
     var deferred = q.defer();
-    xvfb.stop(function() {
-      chromeProcess.kill('SIGINT');
+    xvfb.stop(function () {
+      chromeProcess && chromeProcess.kill('SIGINT');
       deferred.promise.then(allDone);
     });
-  };
+  });
 
   this.forceKill = function () {
-    var self = this;
-
     return q.promise(function (resolve) {
-      self.kill(resolve);
+      self.emit('kill', resolve);
     });
   }
 };
